Remove SpannerKaseDatabaseClientImplementation.
